ScreenShield: show notifications in the locked screen
Rework lock handling in the message tray to separate summary handling
and notification handling. Now the message tray is completely hidden
when the screen is locked, but exceptions can be made for individual
transient notifications.
Non transient sources are shown in the middle of the lock screen. Resident
notifications (such as those from Rhythmbox) are shown in full, while
dpersistent ones are displayed as icon and message count.