StScrollView: add new policy for invisibile but allocated scrollbars
In certain situations, such as the login screen, we need the StScrollView
to have constant width, even if the height (and therefore the need for
vertical scrollbars) changes. To allow that, always request the right size
for scrollbars, and allocate them even if invisible if the policy says so.
This moves away from Gtk.PolicyType, although the new enumeration has the
same values, so it should be backwards compatible.

https://bugzilla.gnome.org/show_bug.cgi?id=686881