Shell: introduce a new ShellShaderEffect class
This is like ClutterShaderEffect (a ClutterEffect for applying arbitrary
GLSL code), except that it exposes and uses the CoglSnippet API instead
of the legacy CoglShader one, so that we don't override Cogl internal
GLSL generation. In particular, it gets us texture lookup for free.
The design is based on the similar API that's exposed by MetaBackgroundActor.

https://bugzilla.gnome.org/show_bug.cgi?id=669798